Associations implemented for sentence generation
We now have associations working for creating sentences. They are essentially worked into the formatting system, so different types of words could end up having slightly different systems for finding associations if wanted.

A few highlights at minimum random unassociated words out of 21 generated:
"a train is like a railroad but if it was fast"
"if you become a beggar then you are legally allowed to plead bum"
"why do we have native indian if we have country?" (oof colonialism tw)
"a fascism is like a rule but if it was evil"
"if you become a hedge then you are legally allowed to trim fence" (doesn't make sense itself but it's funny being backward)
"thesis has big paper energy"
The rest feel more connected than before, even if they don't mean anything. Playing with the percentage of randomness and adding other factors could improve it.